backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
psycopg2-binary==2.9.11
python-dotenv==1.0.0
